requirements.txt
tensorflow-gpu==1.14.0
Keras==2.3.1
numpy==1.16.4
h5py==2.9.0
lxml==4.3.4
mrcfile==1.1.2
scikit-learn==0.22.2.post1
scikit-image==0.15.0
matplotlib==3.1.0
PyQt5==5.13.2
pyqtgraph==0.10.0
openpyxl==3.0.3

